+++ com/hbm/entity/EntityMissileIncendiaryStrong.java
@@ -143,7 +143,7 @@ public class EntityMissileIncendiaryStrong extends EntityMissileBase {
         {
     		if(!this.worldObj.isRemote)
     		{
-    			this.worldObj.newExplosion((Entity)null, (double)((float)this.posX + 0.5F), (double)((float)this.posY + 0.5F), (double)((float)this.posZ + 0.5F), 25.0F, true, true);
+    			this.worldObj.newExplosion((Entity)null, (float)this.posX + 0.5F, (float)this.posY + 0.5F, (float)this.posZ + 0.5F, 25.0F, true, true);
     			ExplosionChaos.flameDeath(this.worldObj, (int)((float)this.posX + 0.5F), (int)((float)this.posY + 0.5F), (int)((float)this.posZ + 0.5F), 25);
     		}
 		this.setDead();
